fix(voice): final VOSK detector and service fixes\n\n- Convert SAMPLE_RATE to Float for Recognizer constructor\n- Rewrite WakeWordService cleanly: fix Context/startForegroundService errors\n- Use proper PendingIntent.getActivity without nullable let\n- Ensure imports are correct (MainActivity, Context)\n- Remove duplicate companion object blocks\n- All compilation errors resolved

+++ app/src/main/kotlin/com/auramusic/app/voice/wakeword/VoskWakeWordDetector.kt
@@ -51,7 +51,7 @@ class VoskWakeWordDetector @Inject constructor(
                 // Download model if not present
                 val modelPath = ensureModel()
                 model = Model(modelPath)
-                recognizer = Recognizer(model, SAMPLE_RATE)
+                recognizer = Recognizer(model, SAMPLE_RATE.toFloat())
 
                 // Start audio loop
                 startAudioRecording()
